requirements.txt
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
flask==2.1.3
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.6
requests==2.31.0
requests-oauthlib==1.1.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.3.7
